refactor(views): replace function views by class views
