+++ apps/core/templates/core/settings/edit.html
@@ -2,7 +2,7 @@
 {% load crispy_forms_tags %}
 
 
-{% block title %}Ajouter un serveur{% endblock title %}
+{% block title %}Modifier le serveur{% endblock title %}
 
 {% block body %}
     <h2>Paramètres globaux</h2>
