+++ apps/core/views/__init__.py
@@ -1,0 +1,10 @@
+from .DefaultView import DefaultView
+from .ServerCreateView import ServerCreateView
+from .ServerDeleteView import ServerDeleteView
+from .ServerEditView import ServerEditView
+from .ServerListView import ServerListView
+from .ServerManageView import ServerManageView
+from .ServerPermissionView import ServerPermissionView
+from .ServerPropertiesView import ServerPropertiesView
+from .ServerStartView import ServerStartView
+from .ServerStopView import ServerStopView
